+++ tests/misc_tests/type_eq_ui_tests/stw_macro_variant_attribute-err.rs
@@ -2,7 +2,13 @@ typewit::simple_type_witness!{
     enum TypeParams {
         #[asdkaspokaopskd]
         U8 = u8,
+        /// Some docs
         #[cfg(any())]
-        U16 = u16 
+        U16 = u16,
+        /// Some docs
+        #[cfg(all())]
+        U16 = u16,
     }
 }
+
+fn main(){}